커뮤니티 글작성시 다시 목록으로 redirect 문제 수정 및 코드 리팩토링
